// F1UI/F1-UI-Angular/src/app/api/model/pointsDto.ts
/**
 * F1Website
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PointsDto { 
    raceId?: number;
    driverId?: number;
    points?: number | null;
    position?: number | null;
    driverName?: string | null;
}